moved components to the right folders
