+++ src/components/MainHeading/MainHeading.tsx
@@ -1,5 +1,4 @@
 'use client';
-import { motion } from 'framer-motion';
 import React from 'react';
 
 export const MainHeading = () => {
