+++ src/components/MainHeading/index.ts
@@ -1,0 +1,1 @@
+export { MainHeading } from './MainHeading';
