+++ src/components/ProjectsLink/index.ts
@@ -1,0 +1,1 @@
+export { ProjectsLink } from './ProjectsLink';
